requirements.txt
requests==2.32.3
prettytable==3.12.0
click==8.1.7
PyYAML==6.0.2
urllib3==1.26.20
pycryptodome==3.21.0